refactor: narrow Score to i16 and depth/ply to u8 throughout search (#18)
Score was i32 and depth/ply were u32/u16, wider than any real value they
ever hold: material plus piece-square terms stay in the low five figures
at the extreme, MATE leaves headroom under i16::MAX, and depth/ply are
bounded by max_depth (64 by default, u8-sized in practice). tt.rs's Entry
already committed to i16/u8 for its own fields, so every narrowing cast at
that boundary (`as i16`, `as u8`) was really just Score/depth/ply being
wider than they needed to be one layer up, not something Entry itself
required.

Carrying the narrower types through the call graph instead removes those
casts entirely rather than relocating them: `Score::from(ply)` and
`Score::from(self.score)` stay plain widening conversions either way now
that both sides are already the same width, and `SearchResult`,
`Response::Info`, and `GoOptions::depth` all follow the same width so
nothing has to cast back and forth at the UCI boundary either. The one
cast this doesn't remove is `search::tt`'s `u64` key/mask into a `Vec`
index: that width is inherent to a 64-bit Zobrist hash, not something
narrowing `Score`/depth/ply touches, so it stays a checked `usize::try_from`.

`Move::bits`'s `#[allow(dead_code)]` is also removed: `Tt::store` calling
it made the allow's own reason stale, not something this refactor
introduced but adjacent enough to fix while touching the same file.

Not fixed here, flagged for follow-up: `Tt::store` stores `key & mask`
(the masked value, identical for every key at a given index) into
`Entry.key` instead of the full `key`, which silently defeats the
collision check `probe` depends on; and `Tt::probe` linearly scans every
entry with a matching masked key instead of indexing directly via
`key & mask`, an O(n) scan where this data structure is supposed to be
O(1). Confirmed both with a temporary diagnostic test (removed before this
commit): a key that was never stored still returns a hit belonging to an
unrelated key colliding at the same index. Also worth another look before
this lands anywhere real: `Tt::new`'s power-of-two rounding
(`next_power_of_two() >> 1`) silently halves the usable table whenever the
byte budget lands on an exact power of two already (confirmed the same
way), since `next_power_of_two` on an already-power-of-two input returns
that same value rather than the next one up.

`uci::session::run` still doesn't own or thread a `Tt` through `Search`;
the `use` and commented-out construction are staged for that, not
connected to the loop yet.

+++ turox-engine/benches/search.rs
@@ -22,7 +22,7 @@ use turox_engine::search::Search;
 const STARTPOS: &str = "rnbqkbnr/pppppppp/8/8/8/8/PPPPPPPP/RNBQKBNR w KQkq - 0 1";
 const KIWIPETE: &str = "r3k2r/p1ppqpb1/bn2pnp1/3PN3/1p2P3/2N2Q1p/PPPBBPPP/R3K2R w KQkq - 0 1";
 
-fn bench_search(c: &mut Criterion, name: &str, fen: &str, depth: u32) {
+fn bench_search(c: &mut Criterion, name: &str, fen: &str, depth: u8) {
     let board = Board::try_from_fen(fen).expect("valid FEN");
     let nodes = Search::new(Vec::new()).search(&board, depth).nodes;
